requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
blas=1.0=mkl
boost=1.70.0=py36h9de70de_1
boost-cpp=1.70.0=h8e57a91_2
bzip2=1.0.8=h516909a_2
ca-certificates=2020.1.1=0
cairo=1.16.0=hfb77d84_1002
certifi=2019.11.28=py36_0
cffi=1.13.2=py36h2e261b9_0
cudatoolkit=10.0.130=0
cycler=0.10.0=pypi_0
decorator=4.4.1=pypi_0
dgl=0.4.2=pypi_0
fontconfig=2.13.1=h86ecdb6_1001
freetype=2.9.1=h8a8886c_1
gettext=0.19.8.1=hc5be6a0_1002
glib=2.58.3=py36h6f030ca_1002
icu=64.2=he1b5a44_1
intel-openmp=2019.4=243
joblib=0.14.1=pypi_0
jpeg=9b=h024ee3a_2
kiwisolver=1.1.0=pypi_0
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libiconv=1.15=h516909a_1005
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_0
libuuid=2.32.1=h14c3975_1000
libxcb=1.13=h14c3975_1002
libxml2=2.9.10=hee79883_0
matplotlib=3.1.2=pypi_0
mkl=2019.4=243
mkl-service=2.3.0=py36he904b0f_0
mkl_fft=1.0.15=py36ha843d7b_0
mkl_random=1.1.0=py36hd6b4f25_0
ncurses=6.1=he6710b0_1
networkx=2.4=pypi_0
ninja=1.9.0=py36hfd86e86_0
numpy=1.18.1=py36h4f9e942_0
numpy-base=1.18.1=py36hde5b4d6_1
olefile=0.46=py36_0
openssl=1.0.2u=h7b6447c_0
pandas=1.0.0=py36hb3f55d8_0
pcre=8.43=he1b5a44_0
pillow=7.0.0=py36hb39fc2d_0
pip=20.0.2=py36_1
pixman=0.38.0=h516909a_1003
pthread-stubs=0.4=h14c3975_1001
pycairo=1.19.0=py36h438ddbb_0
pycparser=2.19=py36_0
pyparsing=2.4.6=pypi_0
python=3.6.3=h6c0c0dc_5
python-dateutil=2.8.1=py_0
pytorch=1.2.0=py3.6_cuda10.0.130_cudnn7.6.2_0
pytz=2019.3=py_0
rdkit=2019.09.3=py36hb31dc5d_0
readline=7.0=h7b6447c_5
scikit-learn=0.22.1=pypi_0
scipy=1.4.1=pypi_0
setuptools=45.1.0=py36_0
six=1.14.0=py36_0
sqlite=3.30.1=h7b6447c_0
tk=8.6.8=hbc83047_0
torchvision=0.4.0=py36_cu100
tqdm=4.42.0=py_0
wheel=0.34.1=py36_0
xorg-kbproto=1.0.7=h14c3975_1002
xorg-libice=1.0.10=h516909a_0
xorg-libsm=1.2.3=h84519dc_1000
xorg-libx11=1.6.9=h516909a_0
xorg-libxau=1.0.9=h14c3975_0
xorg-libxdmcp=1.1.3=h516909a_0
xorg-libxext=1.3.4=h516909a_0
xorg-libxrender=0.9.10=h516909a_1002
xorg-renderproto=0.11.1=h14c3975_1002
xorg-xextproto=7.3.0=h14c3975_1002
xorg-xproto=7.0.31=h14c3975_1007
xz=5.2.4=h14c3975_4
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
